+++ authomatic/__init__.py
@@ -18,4 +18,4 @@ and access **his/her protected resources**.
 
 """
 
-from core import Authomatic, setup, login, provider_id, access, async_access, credentials, request_elements, backend
+from .core import Authomatic, setup, login, provider_id, access, async_access, credentials, request_elements, backend