+++ authomatic/core.py
@@ -4,7 +4,7 @@ from xml.etree import ElementTree
 import collections
 import copy
 import datetime
-import exceptions
+from . import exceptions
 import hashlib
 import hmac
 import logging
@@ -12,8 +12,8 @@ import json
 import pickle
 import threading
 import time
-import urllib
-import urlparse
+import six
+from six.moves.urllib import parse
 import sys
 
 from authomatic.exceptions import SessionError, CredentialsError, RequestElementsError
@@ -40,8 +40,8 @@ def normalize_dict(dict_):
         Normalized dictionary.
     """
 
-    return dict([(k, v[0] if not type(v) is str and len(v) == 1 else v)
-                 for k, v in dict_.items()])
+    return dict([(k, v[0] if not isinstance(v, str) and len(v) == 1 else v)
+                 for k, v in list(dict_.items())])
 
 
 def items_to_dict(items):
@@ -120,8 +120,7 @@ def provider_id():
 
 def escape(s):
     """Escape a URL including any /."""
-    return urllib.quote(s.encode('utf-8'), safe='~')
-
+    return parse.quote(s.encode('utf-8'), safe='~')
 
 def json_qs_parser(body):
     """
@@ -134,7 +133,6 @@ def json_qs_parser(body):
         :class:`dict`, :class:`list` if input is JSON or query string,
         :class:`xml.etree.ElementTree.Element` if XML.
     """
-
     try:
         # Try JSON first.
         return json.loads(body)
@@ -148,7 +146,7 @@ def json_qs_parser(body):
         pass
 
     # Finally query string.
-    return dict(urlparse.parse_qsl(body))
+    return dict(parse.parse_qsl(body))
 
 
 def import_string(import_name, silent=False):
@@ -177,7 +175,7 @@ def resolve_provider_class(class_):
     :param class_name: :class:`string` or :class:`authomatic.providers.BaseProvider` subclass.
     """
 
-    if type(class_) in (str, unicode):
+    if isinstance(class_, str):
         # prepare path for authomatic.providers package
         path = '.'.join([__package__, 'providers', class_])
 
@@ -198,7 +196,7 @@ def id_to_name(config, short_name):
         Value of the id parameter in the :ref:`config` to search for.
     """
 
-    for k, v in config.items():
+    for k, v in list(config.items()):
         if v.get('id') == short_name:
             return k
             break
@@ -236,7 +234,7 @@ class ReprMixin(object):
         # construct keyword arguments
         args = []
 
-        for k, v in self.__dict__.items():
+        for k, v in list(self.__dict__.items()):
 
             # ignore attributes with leading underscores and those listed in _repr_ignore
             if v and not k.startswith('_') and not k in self._repr_ignore:
@@ -312,15 +310,9 @@ class Future(threading.Thread):
 
 
 class Session(object):
-    """
-    A dictionary-like secure cookie session implementation.
-    """
-
-    # List of keys which values are not json serializable.
-    NOT_JSON_SERIALIZABLE = ['_yadis_services__openid_consumer_',
-                             '_openid_consumer_last_token']
-
-    def __init__(self, adapter, secret, name='authomatic', max_age=600, secure=False):
+    """A dictionary-like secure cookie session implementation."""
+    def __init__(self, adapter, secret, name='authomatic', max_age=600,
+                 secure=False):
         """
         :param str secret:
             Session secret used to sign the session cookie.
@@ -329,7 +321,8 @@ class Session(object):
         :param int max_age:
             Maximum allowed age of session cookie nonce in seconds.
         :param bool secure:
-            If ``True`` the session cookie will be saved wit ``Secure`` attribute.
+            If ``True`` the session cookie will be saved with ``Secure``
+            attribute.
         """
 
         self.adapter = adapter
@@ -337,9 +330,7 @@ class Session(object):
         self.secret = secret
         self.max_age = max_age
         self.secure = secure
-
         self._data = {}
-
 
     def create_cookie(self, delete=None):
         """
@@ -349,32 +340,29 @@ class Session(object):
             If ``True`` the cookie value will be ``deleted`` and the
             Expires value will be ``Thu, 01-Jan-1970 00:00:01 GMT``.
         """
-
         value = 'deleted' if delete else self._serialize(self.data)
-
-        split_url = urlparse.urlsplit(self.adapter.url)
+        split_url = parse.urlsplit(self.adapter.url)
         domain = split_url.netloc.split(':')[0]
 
         # Work-around for issue #11, failure of WebKit-based browsers to accept
         # cookies set as part of a redirect response in some circumstances.
-        if not '.' in domain:
+        if '.' not in domain:
             template = '{name}={value}; Path={path}; HttpOnly{secure}{expires}'
         else:
-            template = '{name}={value}; Domain={domain}; Path={path}; HttpOnly{secure}{expires}'
+            template = ('{name}={value}; Domain={domain}; Path={path}; '
+                        'HttpOnly{secure}{expires}')
 
-        return template.format(name=self.name,
-                               value=value,
-                               domain=domain,
-                               path=split_url.path,
-                               secure='; Secure' if self.secure else '',
-                               expires='; Expires=Thu, 01-Jan-1970 00:00:01 GMT' if delete else '')
-
+        return template.format(
+            name=self.name,
+            value=value,
+            domain=domain,
+            path=split_url.path,
+            secure='; Secure' if self.secure else '',
+            expires='; Expires=Thu, 01-Jan-1970 00:00:01 GMT' if delete else ''
+        )
 
     def save(self):
-        """
-        Adds the session cookie to headers.
-        """
-
+        """Adds the session cookie to headers."""
         if self.data:
             # Set the cookie header.
             self.adapter.set_header('Set-Cookie', self.create_cookie())
@@ -382,26 +370,17 @@ class Session(object):
             # Reset data
             self._data = {}
 
-
     def delete(self):
         self.adapter.set_header('Set-Cookie', self.create_cookie(delete=True))
 
-
     def _get_data(self):
-        """
-        Extracts the session data from cookie.
-        """
-
+        """Extracts the session data from cookie."""
         cookie = self.adapter.cookies.get(self.name)
         return self._deserialize(cookie) if cookie else {}
 
-
     @property
     def data(self):
-        """
-        Gets session data lazily.
-        """
-
+        """Gets session data lazily."""
         if not self._data:
             self._data = self._get_data()
         # Always return a dict, even if deserialization returned nothing
@@ -409,23 +388,18 @@ class Session(object):
             self._data = {}
         return self._data
 
-
     def _signature(self, *parts):
         """
         Creates signature for the session.
         """
 
-        signature = hmac.new(self.secret, digestmod=hashlib.sha1)
-        signature.update('|'.join(parts))
+        signature = hmac.new(six.b(self.secret), digestmod=hashlib.sha1)
+        signature.update(six.b('|'.join(parts)))
         return signature.hexdigest()
-
 
     def _serialize(self, value):
         """
         Converts the value to a signed string with timestamp.
-
-        TODO: Check licence!
-        Taken from `webapp2_extras.securecookie <http://webapp-improved.appspot.com/guide/extras.html>`_.
 
         :param value:
             Object to be serialized.
@@ -436,26 +410,16 @@ class Session(object):
 
         data = copy.deepcopy(value)
 
-        # 1. Handle non json serializable objects.
-        for key in self.NOT_JSON_SERIALIZABLE:
-            if key in data.keys():
-                data[key] = pickle.dumps(data[key])
-
-
-        # 2. Serialize
+        # 1. Serialize
         serialized = json.dumps(data)
 
-        # 3. Encode
+        # 2. Encode
         # Percent encoding produces smaller result then urlsafe base64.
-        encoded = urllib.quote(serialized, '')
+        encoded = parse.quote(serialized, '')
 
-        # Create timestamp
+        # 3. Concatenate
         timestamp = str(int(time.time()))
-
-        # Create signature
         signature = self._signature(self.name, encoded, timestamp)
-
-        # 4. Concatenate
         return '|'.join([encoded, timestamp, signature])
 
 
@@ -470,7 +434,7 @@ class Session(object):
             Deserialized object.
         """
 
-        # 4. Split
+        # 3. Split
         encoded, timestamp, signature = value.split('|')
 
         # Verify signature
@@ -481,31 +445,22 @@ class Session(object):
         if int(timestamp) < int(time.time()) - self.max_age:
             return None
 
-        # 3. Decode
-        decoded = urllib.unquote(encoded)
+        # 2. Decode
+        decoded = parse.unquote(encoded)
 
-        # 2. Deserialize
+        # 1. Deserialize
         deserialized = json.loads(decoded)
 
-        # 1. Unpickle non json serializable objects.
-        for key in self.NOT_JSON_SERIALIZABLE:
-            if key in deserialized.keys():
-                deserialized[key] = pickle.loads(str(deserialized[key]))
-
         return deserialized
-
 
     def __setitem__(self, key, value):
         self._data[key] = value
 
-
     def __getitem__(self, key):
         return self.data.__getitem__(key)
 
-
     def __delitem__(self, key):
         return self._data.__delitem__(key)
-
 
     def get(self, key, default=None):
         return self.data.get(key, default)
@@ -860,7 +815,7 @@ class Credentials(ReprMixin):
         concatenated = '\n'.join(stringified)
 
         # Percent encode.
-        return urllib.quote(concatenated, '')
+        return parse.quote(concatenated, '')
 
 
     @classmethod
@@ -879,10 +834,10 @@ class Credentials(ReprMixin):
         """
 
         # Accept both serialized and normal.
-        if type(credentials) is Credentials:
+        if isinstance(credentials, Credentials):
             return credentials
 
-        decoded = urllib.unquote(credentials)
+        decoded = parse.unquote(credentials)
 
         split = decoded.split('\n')
 
@@ -1120,7 +1075,7 @@ class Response(ReprMixin):
         """
 
         if not self._content:
-            self._content = self.httplib_response.read()
+            self._content = self.httplib_response.read().decode('utf-8')
         return self._content
 
 
@@ -1202,7 +1157,7 @@ class RequestElements(tuple):
         Query string of the request.
         """
 
-        return urllib.urlencode(self.params)
+        return parse.urlencode(self.params)
 
     @property
     def full_url(self):