+++ examples/flask/functional_test/main.py
@@ -34,7 +34,7 @@ def login(provider_name):
                               provider_name)
 
     if result:
-        response.data += fixtures.render_login_result(result)
+        response.data += fixtures.render_login_result(result).encode()
 
     return response
 
@@ -43,4 +43,11 @@ if __name__ == '__main__':
     # This does nothing unles you run this module with --testliveserver flag.
     import liveandletdie
     liveandletdie.Flask.wrap(app)
+
+    # Mac
+    # ln -sf /System/Library/Frameworks/Python.framework/Versions/2.6/Extras/lib/python/OpenSSL venv/lib/python2.7/
+    # http://www.oesmith.co.uk/2011/05/23/pyopenssl-in-a-virtualenv-on-osx.html
+
+    # Unix
+    # ???
     app.run(debug=True, port=8080)