+++ tests/functional_tests/__init__.py
@@ -1,1 +1,1 @@
-import fixtures
+from . import fixtures