+++ tests/functional_tests/fixtures/__init__.py
@@ -5,7 +5,8 @@ from collections import namedtuple
 import pkgutil
 import sys
 import time
-
+import six
+from six.moves import reload_module
 from jinja2 import Environment, FileSystemLoader
 
 # Add path of the functional_tests_path package to PYTHONPATH.
@@ -42,8 +43,9 @@ def render_login_result(result):
 
     """
 
-    reload(sys)
-    sys.setdefaultencoding('utf-8')
+    reload_module(sys)
+    if six.PY2:
+        sys.setdefaultencoding('utf-8')
 
     response = None
     original_credentials = {}
@@ -58,7 +60,7 @@ def render_login_result(result):
                 response = result.user.credentials.refresh(force=True)
                 refreshed_credentials.update(result.user.credentials.__dict__)
 
-        user_properties = ASSEMBLED_CONFIG.values()[0]['user'].keys()
+        user_properties = list(ASSEMBLED_CONFIG.values())[0]['user'].keys()
         template = env.get_template('login.html')
         return template.render(result=result,
                                providers=ASSEMBLED_CONFIG.keys(),
@@ -96,18 +98,18 @@ def get_configuration(provider):
     # Add additional class attributes which are not allowed to be passed
     # to the namedtuple
     Res.email_escaped = conf['user_email'].replace('@', '\u0040')
-    Res.no_email = [conf['user_email'], Res.email_escaped, 'email', 'e-mail']
-    Res.no_phone = [conf['user_phone'], 'phone']
     Res.no_birth_date = [conf['user_birth_year'], 'birth']
-    Res.no_gender = [conf['user_gender'], 'gender']
-    Res.no_locale = [conf['user_locale'], 'language', 'locale']
+    Res.no_city = [conf['user_city'], 'city']
+    Res.no_email = [conf['user_email'], Res.email_escaped, 'email', 'e-mail']
     Res.no_first_name = ['"{0}"'.format(conf['user_first_name']), 'first']
     Res.no_last_name = ['"{0}"'.format(conf['user_last_name']), 'last']
+    Res.no_gender = [conf['user_gender'], 'gender']
+    Res.no_locale = [conf['user_locale'], 'language', 'locale']
     Res.no_nickname = ['nickname', conf['user_nickname']]
-    Res.no_username = ['username', '"{}"'.format(conf['user_username'])]
-    Res.no_timezone = ['timezone']
+    Res.no_phone = [conf['user_phone'], 'phone']
     Res.no_postal_code = [conf['user_postal_code'], 'postal', 'zip']
-    Res.no_city = [conf['user_city'], 'city']
+    Res.no_timezone = ['timezone']
+    Res.no_username = ['username', '"{}"'.format(conf['user_username'])]
     Res.no_location = [conf['user_country'], 'city',
         'country', 'location'] + Res.no_postal_code + Res.no_city